refactor ($step3): 패키지명 변경
