+++ src/main/java/step2_3/domain/Lotto.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2_3.domain;
 
 import java.util.List;
