+++ src/test/java/step2_3/LottoIssuerTest.java
@@ -1,7 +1,9 @@
-package step2;
+package step2_3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2_3.domain.Lotto;
+import step2_3.domain.LottoIssuer;
 
 import java.util.List;
 
@@ -19,7 +21,7 @@ public class LottoIssuerTest {
 	@DisplayName(value = "입력받은 갯수만큼 로또를 자동 발급 하는 기능 테스트")
 	@Test
 	void issueAutoLotto() {
-		List<Integer[]> lottoResult = LottoIssuer.issueAutoLottos(14);
+		List<Lotto> lottoResult = LottoIssuer.issueAutoLottos(14);
 		assertThat(lottoResult).hasSize(14);
 	}
